+++ .config/ESLint/eslint.config.js
@@ -3,8 +3,6 @@ import antfu, {
     GLOB_JSX,
 } from "@antfu/eslint-config";
 import tsdoc from "eslint-plugin-tsdoc";
-// @ts-ignore
-import turbo from "eslint-plugin-turbo";
 
 /**
  * @type {import("@antfu/eslint-config").TypedFlatConfigItem['rules']}
@@ -28,7 +26,6 @@ const rules = {
     "perfectionist/sort-array-includes": [
         "warn",
         {
-            "spread-last": true,
         },
     ],
     "perfectionist/sort-exports": [
@@ -38,7 +35,7 @@ const rules = {
     "perfectionist/sort-imports": [
         "warn",
         {
-            "groups": [
+            groups: [
                 "side-effect-style", // import "style.css";
                 "side-effect", // import "module";
 
@@ -89,10 +86,10 @@ const rules = {
                 "style", // import styles from "./index.module.css";
                 "object", // import log = console.log;
             ],
-            "internal-pattern": [
+            internalPattern: [
                 "@/**",
             ],
-            "custom-groups": {
+            customGroups: {
                 value: {
                     $node: "node:**",
                     $repo: "@repo/**",
@@ -115,20 +112,19 @@ const rules = {
     "perfectionist/sort-named-exports": [
         "warn",
         {
-            "group-kind": "values-first",
+            groupKind: "values-first",
         },
     ],
     "perfectionist/sort-named-imports": [
         "warn",
         {
-            "group-kind": "values-first",
-            "ignore-alias": false,
+            groupKind: "values-first",
+            ignoreAlias: false,
         },
     ],
     "perfectionist/sort-union-types": [
         "warn",
         {
-            "nullable-last": true,
         },
     ],
 };
@@ -159,6 +155,7 @@ export default antfu({
     formatters: {
         css: "prettier",
         html: "prettier",
+        xml: "prettier",
         markdown: "dprint",
         graphql: "prettier",
         prettierOptions: {
@@ -297,6 +294,7 @@ export default antfu({
             ],
         },
     },
+    less: {},
     ignores: [
         "./dist",
         "./temp",